+++ src/controllers/course/question.controller.js
@@ -3,7 +3,9 @@ const { Question } = require('../../database/models');
 const QuestionController = {
     async createQuestion(req, res) {
         try {
-            const { content, explanation, numericalOrder, unitId } = req.body;
+            const { courseId, unitId } = req.params;
+            if (isNaN(parseInt(courseId)) || isNaN(parseInt(unitId))) return res.status(400).json({ error: 'Id must be a number' });
+            const { content, explanation, numericalOrder } = req.body;
             const question = await Question.create({ content, explanation, numericalOrder, unitId });
             res.status(201).json(question);
         } catch (error) {
@@ -13,7 +15,11 @@ const QuestionController = {
   
     async getAllQuestions(req, res) {
         try {
-            const questions = await Question.findAll();
+            const { courseId, unitId } = req.params;
+            if (isNaN(parseInt(courseId)) || isNaN(parseInt(unitId))) return res.status(400).json({ error: 'Id must be a number' });
+            const questions = await Question.findAll(
+                { where: { unitId } }
+            );
             res.status(200).json(questions);
         } catch (error) {
             res.status(500).json({ error: error.message });
@@ -22,8 +28,11 @@ const QuestionController = {
   
     async getQuestionById(req, res) {
         try {
-            const { id } = req.params;
-            const question = await Question.findByPk(id);
+            const { courseId, unitId, id } = req.params;
+            if (isNaN(parseInt(courseId)) || isNaN(parseInt(unitId) || isNaN(parseInt(id)))) return res.status(400).json({ error: 'Id must be a number' });
+            const question = await Question.findByPk(id, {
+                where: { courseId, unitId },
+            });
             if (!question) return res.status(404).json({ error: 'Question not found' });
             res.status(200).json(question);
         } catch (error) {
@@ -33,7 +42,8 @@ const QuestionController = {
   
     async updateQuestion(req, res) {
         try {
-            const { id } = req.params;
+            const { courseId, unitId, id } = req.params;
+            if (isNaN(parseInt(courseId)) || isNaN(parseInt(unitId) || isNaN(parseInt(id)))) return res.status(400).json({ error: 'Id must be a number' });
             const { content, explanation, numericalOrder } = req.body;
             const updated = await Question.update(
             { content, explanation, numericalOrder },
@@ -48,7 +58,8 @@ const QuestionController = {
   
     async deleteQuestion(req, res) {
         try {
-            const { id } = req.params;
+            const { courseId, unitId, id } = req.params;
+            if (isNaN(parseInt(courseId)) || isNaN(parseInt(unitId) || isNaN(parseInt(id)))) return res.status(400).json({ error: 'Id must be a number' });
             const deleted = await Question.destroy({ where: { questionId: id } });
             if (!deleted) return res.status(404).json({ error: 'Question not found' });
             res.status(200).json({ message: 'Question deleted successfully' });
