+++ src/controllers/course/unit.controller.js
@@ -3,7 +3,9 @@ const { Unit } = require('../../database/models');
 const UnitController = {
     async createUnit(req, res) {
         try {
-            const { unitName, numericalOrder, courseId } = req.body;
+            const { courseId } = req.params;
+            if (isNaN(parseInt(courseId))) return res.status(400).json({ error: 'Id must be a number' });
+            const { unitName, numericalOrder } = req.body;
             const unit = await Unit.create({ unitName, numericalOrder, courseId });
             res.status(201).json(unit);
         } catch (error) {
@@ -13,9 +15,11 @@ const UnitController = {
   
     async getAllUnits(req, res) {
         try {
-            const units = await Unit.findAll({
-            include: [{ model: Question, as: 'questions' }], // Include related questions
-            });
+            const { courseId } = req.params;
+            if (isNaN(parseInt(courseId))) return res.status(400).json({ error: 'Id must be a number' });
+            const units = await Unit.findAll(
+            { where: { courseId } }
+            );
             res.status(200).json(units);
         } catch (error) {
             res.status(500).json({ error: error.message });
@@ -24,9 +28,10 @@ const UnitController = {
   
     async getUnitById(req, res) {
         try {
-            const { id } = req.params;
+            const { courseId, id } = req.params;
+            if (isNaN(parseInt(courseId)) || isNaN(parseInt(id))) return res.status(400).json({ error: 'Id must be a number' });
             const unit = await Unit.findByPk(id, {
-            include: [{ model: Question, as: 'questions' }],
+            where: { courseId },
             });
             if (!unit) return res.status(404).json({ error: 'Unit not found' });
             res.status(200).json(unit);
@@ -37,7 +42,8 @@ const UnitController = {
   
     async updateUnit(req, res) {
         try {
-            const { id } = req.params;
+            const { courseId, id } = req.params;
+            if (isNaN(parseInt(id)) || isNaN(parseInt(courseId))) return res.status(400).json({ error: 'Id must be a number' });
             const { unitName, numericalOrder } = req.body;
             const updated = await Unit.update(
             { unitName, numericalOrder },
@@ -52,7 +58,8 @@ const UnitController = {
   
     async deleteUnit(req, res) {
         try {
-            const { id } = req.params;
+            const { courseId, id } = req.params;
+            if (isNaN(parseInt(id)) || isNaN(parseInt(courseId))) return res.status(400).json({ error: 'Id must be a number' });
             const deleted = await Unit.destroy({ where: { unitId: id } });
             if (!deleted) return res.status(404).json({ error: 'Unit not found' });
             res.status(200).json({ message: 'Unit deleted successfully' });
